[gitlab] fixing gitlab bug-templates
[skip ci]

The /title <*> construct only works in comments and not when creating
issues, thus it should not be in the templates (but they may be used
as comments to change the title, see GitLab 'Quick Actions'.
The /label ~Upstream adds the Upstream label to the issue regardless
of whether the user adds it or not. Note the label specification
is case sensitive.

Lastly, I could only get the correct drop-down view by having a space in the file name
although I would have preferred to have underscores.

Signed-off-by: Nick Papior <nickpapior@gmail.com>